centralize 'MockDatasetFieldSvc' to remove code duplication #5732
